+++ x/instance/client/cli/query.go
@@ -17,7 +17,7 @@ func GetQueryCmd(queryRoute string, cdc *codec.Codec) *cobra.Command {
 	// Group instance queries under a subcommand
 	instanceQueryCmd := &cobra.Command{
 		Use:                        types.ModuleName,
-		Short:                      fmt.Sprintf("Querying commands for the %s module", types.ModuleName),
+		Short:                      fmt.Sprintf("Query commands for the %s module", types.ModuleName),
 		DisableFlagParsing:         true,
 		SuggestionsMinimumDistance: 2,
 		RunE:                       client.ValidateCmd,
@@ -26,17 +26,18 @@ func GetQueryCmd(queryRoute string, cdc *codec.Codec) *cobra.Command {
 	instanceQueryCmd.AddCommand(
 		flags.GetCommands(
 			GetCmdGet(queryRoute, cdc),
-			GetCmdLists(queryRoute, cdc),
+			GetCmdList(queryRoute, cdc),
 		)...,
 	)
 
 	return instanceQueryCmd
 }
 
+// GetCmdGet implements the get query command.
 func GetCmdGet(queryRoute string, cdc *codec.Codec) *cobra.Command {
 	return &cobra.Command{
-		Use:   "get",
-		Short: "get",
+		Use:   "get [hash]",
+		Short: "Fetch an instance by its hash",
 		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			cliCtx := context.NewCLIContext().WithCodec(cdc)
@@ -53,10 +54,12 @@ func GetCmdGet(queryRoute string, cdc *codec.Codec) *cobra.Command {
 	}
 }
 
-func GetCmdLists(queryRoute string, cdc *codec.Codec) *cobra.Command {
+// GetCmdList implements the list query command.
+func GetCmdList(queryRoute string, cdc *codec.Codec) *cobra.Command {
 	return &cobra.Command{
 		Use:   "list",
-		Short: "list",
+		Short: "Query all the instances",
+		Args:  cobra.NoArgs,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			cliCtx := context.NewCLIContext().WithCodec(cdc)
 			res, _, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/%s", queryRoute, types.QueryList), nil)